+++ chat-widget/src/components/footerstateful/downloadtranscriptstateful/common/defaultLibraryScripts.ts
@@ -1,0 +1,24 @@
+const defaultLibraryScripts = {
+    botframeworkWebChat: {
+        src: "https://cdn.botframework.com/botframework-webchat/4.15.7/webchat.js"
+    },
+    rxJs: {
+        src: "https://cdnjs.cloudflare.com/ajax/libs/rxjs/7.8.0/rxjs.umd.min.js",
+        integrity: "sha512-v0/YVjBcbjLN6scjmmJN+h86koeB7JhY4/2YeyA5l+rTdtKLv0VbDBNJ32rxJpsaW1QGMd1Z16lsLOSGI38Rbg==",
+        crossOrigin: "anonymous",
+        referrerPolicy: "no-referrer"
+    },
+    react: {
+        src: "https://unpkg.com/react@18.2.0/umd/react.production.min.js"
+    },
+    reactDom: {
+        src: "https://unpkg.com/react-dom@18.2.0/umd/react-dom.production.min.js"
+    },
+    markdownIt: {
+        src: "https://cdn.jsdelivr.net/npm/markdown-it@13.0.1/dist/markdown-it.min.js",
+        integrity: "sha256-hNyljag6giCsjv/yKmxK8/VeHzvMDvc5u8AzmRvm1BI=",
+        crossOrigin: "anonymous",
+    }
+};
+
+export default defaultLibraryScripts;
